+++ _templates/component/new/component.ejs.t
@@ -1,0 +1,14 @@
+---
+to: src/components/<%= name %>/<%= name %>.js
+---
+import React from 'react';
+import PropTypes from 'prop-types';
+
+const <%= name %> = (props) => {
+  const {} = props;
+  return <div />;
+};
+
+<%= name %>.propTypes = {};
+
+export default <%= name %>;