+++ _templates/component/new/component_index.ejs.t
@@ -1,0 +1,7 @@
+---
+to: src/components/<%= name %>/index.js
+unless_exists: true
+---
+import <%= name %> from './<%= name %>';
+
+export { <%= name %> };